requirements.txt
pybaseball==2.2.7
polars==0.20.31
pandas==2.2.2
xgboost==2.0.3
ipykernel==6.29.4
pyyaml==6.0.1
hydra-core==1.3.2
scikit-learn==1.5.0
optuna==3.6.1
morphers @ git+https://github.com/play4honor/morphers.git@feature/interface-changes
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.1
lightning==2.3.0
